chatgpt/grpc_qa/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Jack"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
langchain = "^0.0.192"
openai = "^0.27.8"
python-consul = "^1.1.0"
grpcio-tools = "^1.54.2"
faiss-cpu = "^1.7.4"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
